requirements.txt
aioredis==1.3.1
amqp==2.6.0
appdirs==1.4.4
asgi-redis==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
billiard==3.6.3.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-extra-fields==2.0.5
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-framework-base64-fields==0.1.3
djangorestframework==3.11.0
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
kombu==4.6.10
Markdown==3.2.2
msgpack==1.0.0
msgpack-python==0.5.6
Naked==0.1.31
oauthlib==3.1.0
Pillow==7.2.0
pipenv==2020.6.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2020.1
redis==2.10.6
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
shellescape==3.8.1
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
vine==1.3.0
virtualenv==20.0.21
virtualenv-clone==0.5.4
whitenoise==5.1.0
zope.interface==5.1.0
